spending builder test is complete
